+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-server/hadoop-yarn-server-resourcemanager/src/main/java/org/apache/hadoop/yarn/server/resourcemanager/rmnode/RMNode.java
@@ -46,8 +46,8 @@ public interface RMNode {
    */
   public NodeId getNodeID();
 
-  public int getOldestYoungestAge();
-  public void updateOldestYoungestAge(int oldestYoungestAge);
+  public long getOldestYoungestAge();
+  public void updateOldestYoungestAge(long oldestYoungestAge);
   
   /**
    * the hostname of this node
